+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Adding AssetsCache.readBinaryFile
 - Adding support for multi touch tap and drag events
 - Fix animations example
+- Add possibility for infinite and alternating effects
 
 ## 0.20.2
 - Fix text component bug with anchor being applied twice